requirements.txt
torch
torchvision
numpy
opencv-python
mediapipe
trimesh
pygltflib
Pillow
scipy
pytest
